\name{00.setRNG.Intro}
\alias{00.setRNG.Intro}
\docType{package}

\title{setRNG}

\description{Programs to set random number generator (and seed) in R and S.}

\details{ See \code{\link{setRNG-package}} ( in the help system use
   package?setRNG or ?"setRNG-package") for an overview.
}
\keyword{package}


